checkpointing: move to subfolder
