+++ torchft/checkpointing/http_transport.py
@@ -4,13 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""
-Checkpointing
-==============
-
-This module implements methods for checkpointing and resuming training from a checkpoint.
-"""
-
 import io
 import logging
 import socket
@@ -24,68 +17,12 @@ from typing import Generator, Generic, List, Optional, TypeVar
 
 import torch
 
+from torchft.checkpointing.transport import CheckpointTransport
 from torchft.http import _IPv6HTTPServer
 
 logger: logging.Logger = logging.getLogger(__name__)
 
 T = TypeVar("T")
-
-
-class CheckpointTransport(Generic[T], ABC):
-    @abstractmethod
-    def metadata(self) -> str:
-        """
-        Returns a string that will be used by the remote CheckpointTransport to fetch the checkpoint.
-        """
-        ...
-
-    @abstractmethod
-    def send_checkpoint(
-        self, dst_ranks: List[int], step: int, state_dict: T, timeout: timedelta
-    ) -> None:
-        """
-        Sends the checkpoint, only called when there is a rank that is behind.
-
-        This may be async.
-
-        Args:
-            dst_ranks: the ranks to send to
-            step: the step number to send
-            state_dict: the state dict to send
-            timeout: the timeout to wait for the checkpoint to be sent
-        """
-        ...
-
-    def disallow_checkpoint(self) -> None:
-        """
-        Called after send_checkpoint to wait for the checkpoint to be sent.
-
-        Once this returns, the state_dict may be mutated so no further data should be sent.
-        """
-        ...
-
-    @abstractmethod
-    def recv_checkpoint(
-        self, src_rank: int, metadata: str, step: int, timeout: timedelta
-    ) -> T:
-        """
-        Receives the checkpoint from the given rank.
-
-        Args:
-            src_rank: the rank to receive the checkpoint from
-            metadata: the metadata returned by the remote CheckpointTransport
-            step: the step number to receive
-            timeout: the timeout to wait for the checkpoint
-        """
-        ...
-
-    def shutdown(self, wait: bool = True) -> None:
-        """
-        Called to shutdown the checkpoint transport.
-
-        Args:
-            wait: whether to wait for the transport to shutdown
-        """
 
 
 @contextmanager
@@ -107,7 +44,7 @@ def _timed_acquire(
         lock.release()
 
 
-class CheckpointServer(CheckpointTransport[T]):
+class HTTPTransport(CheckpointTransport[T]):
     """
     This is an HTTP server that can be used to transfer checkpoints
     between workers.
